+++ saleor/graphql/order/tests/test_order.py
@@ -126,6 +126,15 @@ def orders(customer_user, channel_USD, channel_PLN):
     )
 
 
+def assert_proper_webhook_called_once(order, status, draft_mock, order_mock):
+    if status == OrderStatus.DRAFT:
+        draft_mock.assert_called_once_with(order)
+        order_mock.assert_not_called()
+    else:
+        draft_mock.assert_not_called()
+        order_mock.assert_called_once_with(order)
+
+
 def test_orderline_query(staff_api_client, permission_manage_orders, fulfilled_order):
     order = fulfilled_order
     query = """
@@ -3331,9 +3340,11 @@ ORDER_LINES_CREATE_MUTATION = """
 
 
 @pytest.mark.parametrize("status", (OrderStatus.DRAFT, OrderStatus.UNCONFIRMED))
+@patch("saleor.plugins.manager.PluginsManager.draft_order_updated")
 @patch("saleor.plugins.manager.PluginsManager.order_updated")
 def test_order_lines_create(
     order_updated_webhook_mock,
+    draft_order_updated_webhook_mock,
     status,
     order_with_lines,
     permission_manage_orders,
@@ -3353,13 +3364,16 @@ def test_order_lines_create(
     # mutation should fail without proper permissions
     response = staff_api_client.post_graphql(query, variables)
     assert_no_permission(response)
+    order_updated_webhook_mock.assert_not_called()
+    draft_order_updated_webhook_mock.assert_not_called()
     assert not OrderEvent.objects.exists()
-    assert not order_updated_webhook_mock.called
 
     # assign permissions
     staff_api_client.user.user_permissions.add(permission_manage_orders)
     response = staff_api_client.post_graphql(query, variables)
-    order_updated_webhook_mock.assert_called_once_with(order)
+    assert_proper_webhook_called_once(
+        order, status, draft_order_updated_webhook_mock, order_updated_webhook_mock
+    )
     assert OrderEvent.objects.count() == 1
     assert OrderEvent.objects.last().type == order_events.OrderEvents.ADDED_PRODUCTS
     content = get_graphql_content(response)
@@ -3377,9 +3391,14 @@ def test_order_lines_create(
     assert data["errors"][0]["variants"] == [variant_id]
 
 
+@patch("saleor.plugins.manager.PluginsManager.draft_order_updated")
 @patch("saleor.plugins.manager.PluginsManager.order_updated")
 def test_order_lines_create_with_unavailable_variant(
-    order_updated_webhook_mock, draft_order, permission_manage_orders, staff_api_client
+    order_updated_webhook_mock,
+    draft_order_updated_webhoook_mock,
+    draft_order,
+    permission_manage_orders,
+    staff_api_client,
 ):
     query = ORDER_LINES_CREATE_MUTATION
     order = draft_order
@@ -3400,13 +3419,16 @@ def test_order_lines_create_with_unavailable_variant(
     assert error["code"] == OrderErrorCode.NOT_AVAILABLE_IN_CHANNEL.name
     assert error["field"] == "input"
     assert error["variants"] == [variant_id]
-    assert not order_updated_webhook_mock.called
+    order_updated_webhook_mock.assert_not_called()
+    draft_order_updated_webhoook_mock.assert_not_called()
 
 
 @pytest.mark.parametrize("status", (OrderStatus.DRAFT, OrderStatus.UNCONFIRMED))
+@patch("saleor.plugins.manager.PluginsManager.draft_order_updated")
 @patch("saleor.plugins.manager.PluginsManager.order_updated")
 def test_order_lines_create_with_existing_variant(
     order_updated_webhook_mock,
+    draft_order_updated_webhook_mock,
     status,
     order_with_lines,
     permission_manage_orders,
@@ -3428,7 +3450,8 @@ def test_order_lines_create_with_existing_variant(
     response = staff_api_client.post_graphql(query, variables)
     assert_no_permission(response)
     assert not OrderEvent.objects.exists()
-    assert not order_updated_webhook_mock.called
+    order_updated_webhook_mock.assert_not_called()
+    draft_order_updated_webhook_mock.assert_not_called()
 
     # assign permissions
     staff_api_client.user.user_permissions.add(permission_manage_orders)
@@ -3439,13 +3462,17 @@ def test_order_lines_create_with_existing_variant(
     data = content["data"]["orderLinesCreate"]
     assert data["orderLines"][0]["productSku"] == variant.sku
     assert data["orderLines"][0]["quantity"] == old_quantity + quantity
-    order_updated_webhook_mock.assert_called_once_with(order)
+    assert_proper_webhook_called_once(
+        order, status, draft_order_updated_webhook_mock, order_updated_webhook_mock
+    )
 
 
 @pytest.mark.parametrize("status", (OrderStatus.DRAFT, OrderStatus.UNCONFIRMED))
+@patch("saleor.plugins.manager.PluginsManager.draft_order_updated")
 @patch("saleor.plugins.manager.PluginsManager.order_updated")
 def test_order_lines_create_with_product_and_variant_not_assigned_to_channel(
     order_updated_webhook_mock,
+    draft_order_updated_webhook_mock,
     status,
     order_with_lines,
     permission_manage_orders,
@@ -3472,13 +3499,16 @@ def test_order_lines_create_with_product_and_variant_not_assigned_to_channel(
     assert error["code"] == OrderErrorCode.PRODUCT_NOT_PUBLISHED.name
     assert error["field"] == "input"
     assert error["variants"] == [variant_id]
-    assert not order_updated_webhook_mock.called
+    order_updated_webhook_mock.assert_not_called()
+    draft_order_updated_webhook_mock.assert_not_called()
 
 
 @pytest.mark.parametrize("status", (OrderStatus.DRAFT, OrderStatus.UNCONFIRMED))
+@patch("saleor.plugins.manager.PluginsManager.draft_order_updated")
 @patch("saleor.plugins.manager.PluginsManager.order_updated")
 def test_order_lines_create_with_variant_not_assigned_to_channel(
     order_update_webhook_mock,
+    draft_order_update_webhook_mock,
     status,
     order_with_lines,
     staff_api_client,
@@ -3508,12 +3538,15 @@ def test_order_lines_create_with_variant_not_assigned_to_channel(
     assert error["code"] == OrderErrorCode.NOT_AVAILABLE_IN_CHANNEL.name
     assert error["field"] == "input"
     assert error["variants"] == [variant_id]
-    assert not order_update_webhook_mock.called
+    order_update_webhook_mock.assert_not_called()
+    draft_order_update_webhook_mock.assert_not_called()
 
 
+@patch("saleor.plugins.manager.PluginsManager.draft_order_updated")
 @patch("saleor.plugins.manager.PluginsManager.order_updated")
 def test_invalid_order_when_creating_lines(
     order_updated_webhook_mock,
+    draft_order_updated_webhook_mock,
     order_with_lines,
     staff_api_client,
     permission_manage_orders,
@@ -3531,7 +3564,8 @@ def test_invalid_order_when_creating_lines(
     content = get_graphql_content(response)
     data = content["data"]["orderLinesCreate"]
     assert data["errors"]
-    assert not order_updated_webhook_mock.called
+    order_updated_webhook_mock.assert_not_called()
+    draft_order_updated_webhook_mock.assert_not_called()
 
 
 ORDER_LINE_UPDATE_MUTATION = """
@@ -3558,9 +3592,11 @@ ORDER_LINE_UPDATE_MUTATION = """
 
 
 @pytest.mark.parametrize("status", (OrderStatus.DRAFT, OrderStatus.UNCONFIRMED))
+@patch("saleor.plugins.manager.PluginsManager.draft_order_updated")
 @patch("saleor.plugins.manager.PluginsManager.order_updated")
 def test_order_line_update(
     order_updated_webhook_mock,
+    draft_order_updated_webhook_mock,
     status,
     order_with_lines,
     permission_manage_orders,
@@ -3586,7 +3622,8 @@ def test_order_line_update(
     # mutation should fail without proper permissions
     response = staff_api_client.post_graphql(query, variables)
     assert_no_permission(response)
-    assert not order_updated_webhook_mock.called
+    order_updated_webhook_mock.assert_not_called()
+    draft_order_updated_webhook_mock.assert_not_called()
 
     # assign permissions
     staff_api_client.user.user_permissions.add(permission_manage_orders)
@@ -3594,8 +3631,9 @@ def test_order_line_update(
     content = get_graphql_content(response)
     data = content["data"]["orderLineUpdate"]
     assert data["orderLine"]["quantity"] == new_quantity
-    order_updated_webhook_mock.assert_called_once_with(order)
-
+    assert_proper_webhook_called_once(
+        order, status, draft_order_updated_webhook_mock, order_updated_webhook_mock
+    )
     removed_items_event = OrderEvent.objects.last()  # type: OrderEvent
     assert removed_items_event.type == order_events.OrderEvents.REMOVED_PRODUCTS
     assert removed_items_event.user == staff_user
@@ -3612,9 +3650,11 @@ def test_order_line_update(
     assert data["errors"][0]["field"] == "quantity"
 
 
+@patch("saleor.plugins.manager.PluginsManager.draft_order_updated")
 @patch("saleor.plugins.manager.PluginsManager.order_updated")
 def test_invalid_order_when_updating_lines(
     order_update_webhook_mock,
+    draft_order_update_webhook_mock,
     order_with_lines,
     staff_api_client,
     permission_manage_orders,
@@ -3631,6 +3671,7 @@ def test_invalid_order_when_updating_lines(
     data = content["data"]["orderLineUpdate"]
     assert data["errors"]
     order_update_webhook_mock.assert_not_called()
+    draft_order_update_webhook_mock.assert_not_called()
 
 
 QUERY_GET_FIRST_EVENT = """
@@ -3742,9 +3783,11 @@ ORDER_LINE_DELETE_MUTATION = """
 
 
 @pytest.mark.parametrize("status", (OrderStatus.DRAFT, OrderStatus.UNCONFIRMED))
+@patch("saleor.plugins.manager.PluginsManager.draft_order_updated")
 @patch("saleor.plugins.manager.PluginsManager.order_updated")
 def test_order_line_remove(
-    order_udate_webhook_mock,
+    order_updated_webhook_mock,
+    draft_order_updated_webhook_mock,
     status,
     order_with_lines,
     permission_manage_orders,
@@ -3767,12 +3810,16 @@ def test_order_line_remove(
     assert OrderEvent.objects.last().type == order_events.OrderEvents.REMOVED_PRODUCTS
     assert data["orderLine"]["id"] == line_id
     assert line not in order.lines.all()
-    assert order_udate_webhook_mock.called_once_with(order)
+    assert_proper_webhook_called_once(
+        order, status, draft_order_updated_webhook_mock, order_updated_webhook_mock
+    )
 
 
+@patch("saleor.plugins.manager.PluginsManager.draft_order_updated")
 @patch("saleor.plugins.manager.PluginsManager.order_updated")
 def test_invalid_order_when_removing_lines(
     order_update_webhook_mock,
+    draft_order_update_webhook_mock,
     staff_api_client,
     order_with_lines,
     permission_manage_orders,
@@ -3789,6 +3836,7 @@ def test_invalid_order_when_removing_lines(
     data = content["data"]["orderLineDelete"]
     assert data["errors"]
     order_update_webhook_mock.assert_not_called()
+    draft_order_update_webhook_mock.assert_not_called()
 
 
 ORDER_UPDATE_MUTATION = """
@@ -3867,7 +3915,7 @@ def test_order_update_with_draft_order(
     error = content["data"]["orderUpdate"]["errors"][0]
     assert error["field"] == "id"
     assert error["code"] == OrderErrorCode.INVALID.name
-    assert order_updated_webhook_mock.called is False
+    order_updated_webhook_mock.assert_not_called()
 
 
 def test_order_update_anonymous_user_no_user_email(
@@ -4041,7 +4089,7 @@ def test_order_add_note_fail_on_empty_message(
     data = content["data"]["orderAddNote"]
     assert data["errors"][0]["field"] == "message"
     assert data["errors"][0]["code"] == OrderErrorCode.REQUIRED.name
-    assert not order_updated_webhook_mock.called
+    order_updated_webhook_mock.assert_not_called()
 
 
 MUTATION_ORDER_CANCEL = """